fs: btrfs: Use vfs_time accessors
This is in preparation for changing VFS inode timestamps to
use 64 bit time.
The VFS inode timestamps are not y2038 safe as they use
struct timespec. These will be changed to use struct timespec64
instead and that is y2038 safe.
But, since the above data type conversion will break the
end file systems, use vfs_time functions to access inode times.

Note that the btrfs_update_time() has not been changed in this
patch. This will need to be updated along with vfs time
data type changes. This is because the function signature is
decided by the VFS layer since it is a VFS inode operations
callback.

Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>