fs: afs: Replace get_seconds() by current_fs_time_sec()
get_seconds() is not appropriate for filesystems as it
doesn't use the right granularity for filesystem timestamps.
Use current_fs_time_sec() instead which will be extended to
perform range checks and clamping on the timestamps.

get_seconds() is also not y2038 safe. current_fs_time_sec()
will be transitioned to use 64 bit time along with vfs in a
separate patch. As part of that effort current_fs_time_sec()
will also be extended to do range checks. Hence, it is
necessary for all file system timestamps to use
current_fs_time* functions.

Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>
Cc: David Howells <dhowells@redhat.com>
Cc: linux-afs@lists.infradead.org