fs: coda: Replace CURRENT_TIME_SEC by current_fs_time_sec()
CURRENT_TIME_SEC is not appropriate for filesystems as it
doesn't check range on filesystem timestamps.
Use current_fs_time_sec() instead which will be extended to
perform range checks and clamping on the timestamps.

CURRENT_TIME_SEC is also not y2038 safe. current_fs_time_sec()
will be transitioned to use 64 bit time along with vfs in a
separate patch.

Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>
Cc: Jan Harkes <jaharkes@cs.cmu.edu>
Cc: coda@cs.cmu.edu
Cc: codalist@coda.cs.cmu.edu