fs: ocfs2: Use timespec64 to represent orphan scan times
struct timespec is not y2038 safe.
All timespec references in the kernel are going to be
replaced with struct timespec64.
Use timespec64 which is y2038 safe to represent orphan
scan times.

Also use appropriate time functions that return time in
timespec64 format. Time functions now return monotonic
time instead of real time as only delta scan times are
relevant and these values are not persistent across
reboots.

The format string for the debug print is still using long
as this is only the time elapsed since the last scan and
long is sufficient to represent this value.

Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>
Cc: Mark Fasheh <mfasheh@suse.com>
Cc: Joel Becker <jlbec@evilplan.org>
Cc: ocfs2-devel@oss.oracle.com