fix get_timespec64() for y2038 safe compat interfaces
get/put_timespec64() interfaces will eventually be used for
conversions between the new y2038 safe struct __kernel_timespec
and struct timespec64.

The new y2038 safe syscalls have a common entry for native
and compat interfaces.
On compat interfaces, the high order bits of nanoseconds are
should to be zeroed out. This is because the application code
or the libc do not garuntee zeroing of these. If used without
zeroing, kernel might be at risk of using timespec values
incorrectly.

Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>