input_event: Provide override for sparc64
The usec part of the timeval is defined as
__kernel_suseconds_t	tv_usec; /* microseconds */

Arnd noticed that sparc64 is the only architecture
that defines __kernel_susecond_t as int rather than long.

This breaks the current y2038 fix for kernel as we only
access and define the timeval struct for non-kernel use cases.
But, there is an another typo in the use of __KERNEL__
qualifier. So this is not yet a bug.

The patch fixes the typo and provides an override for sparc64.

Fixes: 152194fe9c3f ("Input: extend usable life of event timestamps to 2106 on 32 bit systems")
Reported-by: Arnd Bergmann <arnd@arndb.de>
Signed-off-by: Deepa Dinamani <deepa.kernel@gmail.com>